+++ PostService/src/main/java/com/somnus/server/postservice/articles/repository/ArticleRepository.java
@@ -1,0 +1,20 @@
+package com.somnus.server.postservice.articles.repository;
+
+import com.somnus.server.postservice.articles.domain.Article;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.stereotype.Repository;
+
+import javax.transaction.Transactional;
+import java.util.List;
+
+@Repository
+@Transactional
+public interface ArticleRepository extends JpaRepository<Article, Integer> {
+
+    @Query("SELECT a FROM Article a WHERE a.topic = :topic")
+    List<Article> getArticlesMatchingTopic(String topic);
+
+    @Query("SELECT a FROM Article a WHERE a.postName LIKE :nameContent")
+    List<Article> getAllArticlesContainingSubstring(String nameContent);
+}