+++ PostService/src/main/java/com/somnus/server/postservice/notifications/config/PusherInfo.java
@@ -1,0 +1,10 @@
+package com.somnus.server.postservice.notifications.config;
+
+public class PusherInfo {
+
+    // Pusher channel names
+    public static final String ARTICLE_CHANNEL = "article";
+
+    // Pusher Event names
+    public static final String COMMENT_PUBLISHED_EVENT = "comment-published";
+}