+++ PostService/src/test/java/com/somnus/server/postservice/PostserviceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.somnus.server.postservice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class PostserviceApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}