+++ SomnusBackend/src/main/java/com/somnus/server/backend/SomnusBackendApplication.java
@@ -1,6 +1,5 @@
 package com.somnus.server.backend;
 
-import com.somnus.server.backend.notifications.config.PusherConfig;
 import com.somnus.server.backend.teammembers.ContributionService;
 import com.somnus.server.backend.teammembers.TeamMemberService;
 import org.springframework.beans.factory.InitializingBean;
@@ -30,9 +29,6 @@ public class SomnusBackendApplication extends SpringBootServletInitializer imple
 	@Autowired
 	private TeamMemberService teamMemberService;
 
-	@Autowired
-	private PusherConfig pusherConfig;
-
 	public static void main(String[] args) {
 		SpringApplication.run(SomnusBackendApplication.class, args);
 	}
@@ -41,7 +37,6 @@ public class SomnusBackendApplication extends SpringBootServletInitializer imple
 	public void afterPropertiesSet() {
 		contributionService.updateContributionRepo();
 		teamMemberService.updateTeamMemberOldestCommit();
-		pusherConfig.createPusherConnection();
 	}
 
 	@Override